+++ HISTORY.md
@@ -1,11 +1,14 @@
-unreleased
-==========
+1.8.0 / 2025-02-10
+==================
+
   * Use `res.headersSent` when available
+  * Replace `_implicitHeader` with `writeHead` property
   * add brotli support for versions of node that support it
   * Add the enforceEncoding option for requests without `Accept-Encoding` header
 
 1.7.5 / 2024-10-31
-==========
+==================
+
   * deps: Replace accepts with negotiator@~0.6.4
     - Add preference option
   * deps: bytes@3.1.2